+++ src/Components/Authentication/Desktop/ModalManager.tsx
@@ -1,10 +1,9 @@
+// @ts-ignore
 import React, { Component } from "react"
 import { FormikProps } from "formik"
 import { DesktopModal } from "./Components/DesktopModal"
 import { FormSwitcher } from "./FormSwitcher"
 import { handleSubmit as defaultHandleSubmit } from "Components/Authentication/helpers"
-import { track } from "../../../Utils/track"
-import Events from "../../../Utils/Events"
 import {
   InputValues,
   SubmitHandler,
@@ -30,7 +29,6 @@ export interface ModalManagerState {
   copy?: string | null
 }
 
-@track({}, { dispatch: data => Events.postEvent(data) })
 export class ModalManager extends Component<
   ModalManagerProps,
   ModalManagerState
@@ -41,47 +39,15 @@ export class ModalManager extends Component<
   }
 
   openModal = (options: ModalOptions) => {
-    const {
-      mode,
-      contextModule,
-      copy,
-      signupIntent,
-      redirectUrl,
-      destination,
-    } = options
-    console.log("Im in here..")
+    const { mode, copy } = options
+
     this.setState({
       currentType: mode,
       copy,
     })
-
-    // Analytics
-    const event = Object.assign(
-      {
-        action: "Auth impression",
-        type: mode,
-        label: contextModule,
-        modal_copy: copy,
-      },
-      options.mode === "signup"
-        ? {
-            signup_intent: signupIntent,
-            signup_redirect: redirectUrl || destination,
-            onboarding: !redirectUrl,
-          }
-        : null
-    )
-    this.props.tracking.trackEvent(event)
   }
 
   closeModal = () => {
-    this.props.tracking.trackEvent({
-      action: "Click",
-      type: "dismiss",
-      label: "dismiss auth modal",
-      flow: "auth",
-    })
-
     this.setState({
       currentType: null,
     })
